src/containers/human-image-segmentation/model/requirements.txt
numpy
Pillow
onnxruntime-gpu==1.6
boto3